requirements.txt
fastapi
uvicorn
numba
numpy
torch
tqdm
more-itertools
tiktoken==0.3.3
azure-storage-blob
modal-client
